fix: #91 Bottom Tab Bar を画面下部に常時固定 (position: fixed)
スクロール先まで流れて見えなくなる問題を修正。タブバーを
fixed bottom-0 left-0 right-0 z-40 で画面下部に常時固定する。

- shared/lib/useBottomTabBarVisible.ts に表示判定 composable を新設
  (App.vue の root padding-bottom と BottomTabBar 自身の v-if で同じロジックを共有)
- BottomTabBar を fixed position に変更、iOS safe-area-inset-bottom 対応
- App.vue でタブバー表示時は root に padding-bottom: 80px を確保
- AppFooter はタブバー表示時には非表示 (重畳・スペース消費を回避)
